Move buttons to separate package
